--- blocks/script.js
@@ -1,1 +1,0 @@
-import "./common.blocks/modal/modal.js";
