--- create/blocks/script.js
@@ -1,15 +1,0 @@
-"use strict";
-let pwdInp1 = document.querySelector('#pwd');
-let pwdInp2 = document.querySelector('#pwd2');
-let btn = document.querySelector('#btn');
-let ul = document.querySelector('.signup__list');
-let dateBirth = document.querySelector("#dateBirth");
-
-let letter = document.querySelector("#lowLetter");
-let capital = document.querySelector("#capLetter");
-let number = document.querySelector("#number");
-let length = document.querySelector("#length");
-let message = document.querySelector("#message");
-
-import "./signup/__input/signup__input.js";
-import "./signup/__submit/signup__submit.js";
