+++ create/blocks/signup/__error/signup__error.scss
@@ -1,0 +1,6 @@
+.signup {
+    &__error {
+        color: red;
+        margin: -10px 0 15px 2px;
+    }
+}
