+++ create/blocks/signup/__icon/_right/signup__icon_right.scss
@@ -1,0 +1,16 @@
+.signup {
+    &__icon {
+        &_right {
+            cursor: pointer;
+            position: absolute;
+            right: 8px;
+            bottom: 6px;
+            width: 25px;
+            height: 25px;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            transform: scale(1.4);
+        }
+    }
+}
