+++ create/blocks/signup/__icon/signup__icon.scss
@@ -1,0 +1,2 @@
+@import "./_left/signup__icon_left.scss";
+@import "./_right/signup__icon_right.scss";
