+++ create/blocks/signup/__input/signup__input.scss
@@ -18,8 +18,7 @@
     }
    
 }
-
-
+/*
 .fa-eye, .fa-eye-slash,.fa-sort-desc {
     cursor: pointer;
     position: absolute;
@@ -32,9 +31,4 @@
     align-items: center;
     transform: scale(1.4);
 }
-
-.fa-at, .fa-user, .fa-lock {
-    position: absolute;
-    left: 10px;
-    bottom: 13px;
-}
+*/
