+++ create/blocks/signup/__submit/signup__submit.js
@@ -1,3 +1,9 @@
+"use strict";
+let btn = document.querySelector('#btn');
+
+let dateBirth = document.querySelector("#dateBirth");
+
+
 btn.onclick = (event) => {
     if ((pwdInp2.value !== pwdInp1.value)) {
         event.preventDefault();
@@ -5,22 +11,23 @@ btn.onclick = (event) => {
         btn.disabled = true;
         btn.classList.add('signup__submit_disabled');
 
-        let div = `<div id="error" style="color: red;margin: -10px 0 15px 2px;"> Разные пароли. Проверьте пароль </div>`;
+        let div = `<div id="error" class="signup__error"> Разные пароли. Проверьте пароль </div>`;
         dateBirth.insertAdjacentHTML('beforebegin', div);       
 
         pwdInp2.onkeyup = () => {
+            let error = document.querySelector('#error');
             if (pwdInp2.value == pwdInp1.value) {
                 pwdInp2.style.border = '1px solid green';
                 btn.disabled = false;
                 btn.classList.remove('signup__submit_disabled');
                 btn.classList.add('signup__submit_enable');
-                document.querySelector('#error').hidden = true;
+                error.hidden = true;
             } else {
                 pwdInp2.style.border = '1px solid red';
                 btn.disabled = true;
                 btn.classList.remove('signup__submit_enable');
                 btn.classList.add('signup__submit_disabled');
-                document.querySelector('#error').hidden = false;
+                error.hidden = false;
 
             }
         }
