+++ create/blocks/signup/signup.scss
@@ -1,4 +1,5 @@
 @import "./__title/signup__title.scss";
+@import "./__icon/signup__icon.scss";
 @import "./__name-field/signup__name-field.scss";
 @import "./__list/signup__list.scss";
 @import "./__list-item/signup__list-item.scss";
@@ -11,3 +12,4 @@
 @import "./__submit/signup__submit.scss";
 @import "./__explanation/signup__explanation.scss";
 @import "./__agree/signup__agree.scss";
+@import "./__error/signup__error.scss";
