+++ create/blocks/style.scss
