+++ create/index.html
@@ -55,7 +55,7 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
     />
-    <link rel="stylesheet" href="blocks/create.css" />
+    <link rel="stylesheet" href="blocks/style.css" />
   </head>
   <body>
     <header class="header">
@@ -105,7 +105,7 @@
                     > <h2 class="signup__name-field">Имя пользователя:</h2>
                     <strong> <abbr title="required">*</abbr></strong>
                   </label>
-                  <i class="fa fa-user icon"></i>
+                  <i class="signup__icon_left fa fa-user"></i>
                   <input
                     maxlength="20"
                     class="signup__input"
@@ -120,7 +120,7 @@
                     > <h2 class="signup__name-field">адрес email:</h2>
                     <strong> <abbr title="required">*</abbr></strong>
                   </label>
-                  <i class="fa fa-at"></i>
+                  <i class="signup__icon_left fa fa-at"></i>
                   <input
                     class="signup__input"
                     id="email"
@@ -134,7 +134,7 @@
                     > <h2 class="signup__name-field">Пароль:</h2>
                     <strong> <abbr title="required">*</abbr></strong>
                   </label>
-                  <i class="fa fa-lock"></i>
+                  <i class="signup__icon_left fa fa-lock"></i>
                   <input
                     pattern="(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{8,}"
                     class="signup__input"
@@ -143,7 +143,7 @@
                     placeholder="Введите пароль"
                     required
                   />
-                  <i id="eye1" class="signup__icon-eye fa fa-eye"></i>
+                  <i id="eye1" class="signup__icon-eye signup__icon_right fa fa-eye"></i>
                 </li>
                 <ol
                   class="signup__message signup__message_invisible"
@@ -182,7 +182,7 @@
                     > <h2 class="signup__name-field">Повторите пароль:</h2>
                     <strong> <abbr title="required">*</abbr></strong>
                   </label>
-                  <i class="fa fa-lock"></i>
+                  <i class="signup__icon_left fa fa-lock"></i>
                   <input
                     class="signup__input"
                     id="pwd2"
@@ -190,7 +190,7 @@
                     placeholder="Введите пароль повторно"
                     required
                   />
-                  <i id="eye2" class="signup__icon-eye fa fa-eye"></i>
+                  <i id="eye2" class="signup__icon-eye signup__icon_right fa fa-eye"></i>
                 </li>
                 <li class="signup__list-item">
                   <label id="dateBirth" class="signup__label" for="date"> <h2 class="signup__name-field">Дата рождения:</h2> </label>
@@ -208,7 +208,7 @@
                     <option value="/male/">Мужской</option>
                     <option value="/female/">Женский</option>
                   </select>
-                  <i class="fa fa-sort-desc"></i>
+                  <i class="signup__icon_right fa fa-sort-desc"></i>
                 </li>
               </ul>
               <div class="signup__explanation">              
@@ -381,6 +381,7 @@
         </div>
       </div>
     </footer>
-    <script type="module" src="./blocks/script.js"></script>
+    <script src="./blocks/signup/__input/signup__input.js"></script>
+    <script src="./blocks/signup/__submit/signup__submit.js"></script>
   </body>
 </html>
