+++ index.html
@@ -756,6 +756,6 @@
         </div>
       </div>
     </footer>
-    <script type="module" src="blocks/script.js"></script>
+    <script src="./blocks/common.blocks/modal/modal.js"></script>
   </body>
 </html>
