+++ pushkin/blocks/style.css
@@ -16,12 +16,19 @@ html, body {
   height: 100vh;
   overflow: auto;
   background-color: rgba(0, 0, 0, 0.5);
-  padding-top: 50px;
+  padding-top: 15px;
+  /*&_visible {
+      display: block;
+  }
+
+  &_invisible {
+      display: none;
+  }*/
 }
 
 .form {
   position: relative;
-  padding: 30px 0 0 0;
+  padding: 15px 0 0 0;
   border-radius: 10px;
   background-color: #fefefe;
   margin: 0 auto;
@@ -278,6 +285,30 @@ html, body {
   width: 128px;
 }
 
+.actions__signup {
+  width: 100px;
+  margin-right: 10px;
+  box-sizing: border-box;
+  height: 40px;
+  background-color: #ffffff;
+  border: 1px solid #e0e0e0;
+  padding: 0;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.actions__signup-link {
+  list-style-type: none;
+  padding: 5px;
+  text-decoration: none;
+  font-family: "Roboto", sans-serif;
+  font-weight: 500;
+  font-size: 0.8125rem;
+  color: #1d1d1d;
+}
+
 .actions__sign-in {
   box-sizing: border-box;
   width: 78px;
@@ -289,6 +320,7 @@ html, body {
   display: flex;
   align-items: center;
   justify-content: center;
+  margin-right: 10px;
 }
 
 .actions__sign-in-text {
@@ -306,7 +338,7 @@ html, body {
   justify-content: center;
 }
 
-.actions__sign-in:hover {
+.actions__sign-in:hover, .actions__signup:hover {
   animation-name: background-color;
   animation-duration: 500ms;
   animation-fill-mode: forwards;
@@ -316,7 +348,7 @@ html, body {
     background-color: #1d1d1d;
   }
 }
-.actions__sign-in-text:hover {
+.actions__sign-in-text:hover, .actions__signup-link:hover {
   animation-name: color;
   animation-duration: 500ms;
   animation-fill-mode: forwards;
@@ -1055,4 +1087,4 @@ html, body {
   }
 }
 
-/*# sourceMappingURL=blocks.css.map */
+/*# sourceMappingURL=style.css.map */
