+++ pushkin/blocks/style.scss
