+++ pushkin/pushkin.html
@@ -10,7 +10,7 @@
       type="text/css"
       href="https://necolas.github.io/normalize.css/8.0.1/normalize.css"
     />
-    <link rel="stylesheet" href="blocks/blocks.css" />
+    <link rel="stylesheet" href="blocks/style.css" />
     <link
       rel="stylesheet"
       href="https://unpkg.com/flexboxgrid2@7.2.1/flexboxgrid2.css"
@@ -509,6 +509,6 @@
         </div>
       </div>
     </footer>
-    <script type="module" src="blocks/pushkin.js"></script>
+    <script type="module" src="../blocks/common.blocks/modal/modal.js"></script>
   </body>
 </html>
